requirements.txt
Flask==2.3.2
Flask-Cors==3.0.10
psycopg2-binary==2.9.6
neo4j==5.9.0